refactoring mmcv.images
