+++ mmcv/image/__init__.py
@@ -1,18 +1,19 @@
 # Copyright (c) Open-MMLab. All rights reserved.
 from .colorspace import (bgr2gray, bgr2hls, bgr2hsv, bgr2rgb, gray2bgr,
-                         gray2rgb, hls2bgr, hsv2bgr, iminvert, posterize,
-                         rgb2bgr, rgb2gray, solarize)
-from .geometry import (imcrop, imflip, imflip_, impad, impad_to_multiple,
-                       imrotate)
+                         gray2rgb, hls2bgr, hsv2bgr, imconvert, rgb2bgr,
+                         rgb2gray)
+from .geometric import (imcrop, imflip, imflip_, impad, impad_to_multiple,
+                        imrescale, imresize, imresize_like, imrotate,
+                        rescale_size)
 from .io import imfrombytes, imread, imwrite, supported_backends, use_backend
-from .normalize import imdenormalize, imnormalize, imnormalize_
-from .resize import imrescale, imresize, imresize_like, rescale_size
+from .photometric import (imdenormalize, iminvert, imnormalize, imnormalize_,
+                          posterize, solarize)
 
 __all__ = [
-    'solarize', 'posterize', 'imread', 'imwrite', 'imfrombytes', 'bgr2gray',
-    'rgb2gray', 'gray2bgr', 'gray2rgb', 'bgr2rgb', 'rgb2bgr', 'bgr2hsv',
-    'hsv2bgr', 'bgr2hls', 'hls2bgr', 'iminvert', 'imflip', 'imflip_',
-    'imrotate', 'imcrop', 'impad', 'impad_to_multiple', 'imnormalize',
-    'imnormalize_', 'imdenormalize', 'imresize', 'imresize_like', 'imrescale',
-    'use_backend', 'supported_backends', 'rescale_size'
+    'bgr2gray', 'bgr2hls', 'bgr2hsv', 'bgr2rgb', 'gray2bgr', 'gray2rgb',
+    'hls2bgr', 'hsv2bgr', 'imconvert', 'rgb2bgr', 'rgb2gray', 'imrescale',
+    'imresize', 'imresize_like', 'rescale_size', 'imcrop', 'imflip', 'imflip_',
+    'impad', 'impad_to_multiple', 'imrotate', 'imfrombytes', 'imread',
+    'imwrite', 'supported_backends', 'use_backend', 'imdenormalize',
+    'imnormalize', 'imnormalize_', 'iminvert', 'posterize', 'solarize'
 ]
